+++ integration/home_roomcode_test.go
@@ -1,4 +1,4 @@
-package rodtests
+package integration
 
 import (
 	"testing"
